refactor(merkle): move CompactMmr64 to legacy module disabled by default
