+++ crates/merkle/src/lib.rs
@@ -35,17 +35,26 @@ use criterion as _;
 
 #[cfg(feature = "codec")]
 mod codec_impl;
-pub mod error;
-pub mod hasher;
-pub mod mmr;
-pub mod proof;
-pub mod tree;
 
+#[cfg(any(test, feature = "legacy_compact"))]
+mod legacy_compact_mmr;
+
+mod error;
 mod ext;
+mod hasher;
+mod mmr;
+mod proof;
 mod traits;
+mod tree;
 
-pub use ext::Mmr;
-pub use traits::MmrState;
+pub use error::*;
+pub use ext::*;
+pub use hasher::*;
+#[cfg(any(test, feature = "legacy_compact"))]
+pub use legacy_compact_mmr::*;
+pub use proof::*;
+pub use traits::*;
+pub use tree::*;
 
 // Include SSZ-generated types
 #[cfg(feature = "ssz")]
@@ -60,7 +69,6 @@ mod ssz_generated {
     include!(concat!(env!("OUT_DIR"), "/generated_ssz.rs"));
 }
 
-use hasher::{DigestMerkleHasher, DigestMerkleHasherNoPrefix};
 use sha2::Sha256;
 
 /// Merkle hash impl for SHA-256 `Digest` impl.
@@ -71,13 +79,8 @@ pub type Sha256NoPrefixHasher = DigestMerkleHasherNoPrefix<Sha256, 32>;
 
 /// Compatibility alias for the primary hasher trait used across Merkle data structures.
 pub use hasher::MerkleHasher as StrataMerkle;
-// Common re-exports for ergonomic access at the crate root.
-pub use hasher::{MerkleHash, MerkleHasher};
-pub use mmr::CompactMmr64;
-pub use proof::{MerkleProof, ProofData, ProofDataMut, RawMerkleProof};
 // Re-export SSZ-generated concrete types (32-byte hash versions)
 #[cfg(feature = "ssz")]
 pub use ssz_generated::ssz::mmr::*;
 #[cfg(feature = "ssz")]
 pub use ssz_generated::ssz::proof::*;
-pub use tree::BinaryMerkleTree;
